backend/pyproject.toml
[project]
name = "petool-backend"
version = "0.1.0"
description = "Petool AI 中转后端服务"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "openai>=1.0",
    "tiktoken>=0.9",
    "motor>=3.7",
    "pymongo>=4.10",
    "python-jose[cryptography]>=3.3",
    "argon2-cffi>=23.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0",
    "alipay-sdk-python>=3.0",
    "wechatpayv3>=1.3",
    "pydantic[email]>=2.12.5",
]
